Move book trip methods from index.js to domUpdates
